Listen for 'exit' instead of 'close' event
Amends `rl.spawnAsyncProcess()` to listen for a spawned child process's
'exit' event instead of the 'close' event.

The 'close' event is emitted when the stdio streams of a child process
have all terminated, in contrast to 'exit' which is emitted when the
child process ends. A child process's stdio streams can remain open
after it exits if its streams are shared with another process; then,
the 'close' event would not be called when the child process ends.